Hud states u32 -> u8

+++ src/xrGame/cta_game_artefact.h
@@ -21,8 +21,8 @@ public:
 	virtual				~CtaGameArtefact			();
 
 	virtual		bool	Action						(s32 cmd, u32 flags);
-	virtual		void	OnStateSwitch				(u32 S);
-	virtual		void	OnAnimationEnd				(u32 state);
+	virtual		void	OnStateSwitch				(u8 S);
+	virtual		void	OnAnimationEnd				(u8 state);
 	virtual		void	UpdateCLChild				();
 	virtual		bool	CanTake						() const;
 	//virtual		void	net_Export					(NET_Packet& P);
